Test customize feature of template for optimizing website purpose
I create some files to test the customize feature of template. In other words, I check whether or not I can choose which files to load on a particular page. So if a page does not a CSS file, I will not load it to the page. This is for optimization of website.